add documents support

+++ matrix_vision/matrix_vision.py
@@ -102,7 +102,6 @@ class MatrixVision:
             self.draw()
             self.images.append(self.screen.copy())
             counter += 1
-            self.clock.tick(30)
         return self.generate_animation(out_name)
 
     def generate_animation(self, out_name=None):
@@ -124,6 +123,7 @@ class MatrixVision:
         with open(out_name, 'rb') as animation_file:             # so WA is to save to tmp file and then read to io.BytesIO and delete it afterwards
             animation = BytesIO(animation_file.read())
             animation.name = out_name
-        os.remove(out_name)
-        log.info(f"Removed temporary file {out_name}")
+        if not self.save_result:
+            os.remove(out_name)
+            log.info(f"Removed temporary file {out_name}")
         return animation
